Changes to make Squawk smaller: - Move API printing to the romizer - made filed in Klass smaller - Added more details to VMThread.printState() - Removed a couple of virtual methods that snuck into the VM class. - Exclude VM.lookupNative
git-svn-id: file:///home/tomatsu/sandbox/tmp/squawk-mirror/trunk@629 5e201862-2652-49f7-a74a-6bc36218013a

+++ romizer/src/com/sun/squawk/ObjectGraphLoader.java
@@ -104,7 +104,7 @@ public class ObjectGraphLoader {
         UWord oopMapWord = NativeUnsafe.getUWord(address, FieldOffsets.decodeOffset(FieldOffsets.com_sun_squawk_Klass$oopMapWord));
         UWord[] dataMap = getUWordsAt(NativeUnsafe.getAddress(address, FieldOffsets.decodeOffset(FieldOffsets.com_sun_squawk_Klass$dataMap)));
         UWord dataMapWord = NativeUnsafe.getUWord(address, FieldOffsets.decodeOffset(FieldOffsets.com_sun_squawk_Klass$dataMapWord));
-        int dataMapLength = NativeUnsafe.getInt(address, FieldOffsets.decodeOffset(FieldOffsets.com_sun_squawk_Klass$dataMapLength));
+        short dataMapLength = (short)NativeUnsafe.getShort(address, FieldOffsets.decodeOffset(FieldOffsets.com_sun_squawk_Klass$dataMapLength));
         int modifiers = NativeUnsafe.getInt(address, FieldOffsets.decodeOffset(FieldOffsets.com_sun_squawk_Klass$modifiers));
         byte state = (byte) NativeUnsafe.getByte(address, FieldOffsets.decodeOffset(FieldOffsets.com_sun_squawk_Klass$state));
         short instanceSizeBytes = (short) NativeUnsafe.getShort(address, FieldOffsets.decodeOffset(FieldOffsets.com_sun_squawk_Klass$instanceSizeBytes));
